requirements.txt
# Core dependencies
python-telegram-bot==20.7

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Optional: for environment variables
python-dotenv==1.0.0
